Issue-10 working solution
