+++ go.mod
@@ -6,4 +6,5 @@ require (
 	github.com/GuiaBolso/darwin v0.0.0-20191218124601-fd6d2aa3d244
 	github.com/jmoiron/sqlx v1.2.0
 	github.com/lib/pq v1.9.0
+	github.com/pkg/errors v0.9.1
 )
